+++ NEWS.md
@@ -1,3 +1,7 @@
+# eclectic (0.0.0.9001)
+
+* new function: rename_duplicates
+
 # eclectic (development version 0.0.0.9000)
 
 * Package created.